src/app/pyproject.toml
[project]
name = "gainwell-genie-app"
dynamic = ["version"]
description = "Gainwell Genie Data App: ask questions in natural language, view data, download and visualize."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.37.0",
    "databricks-sdk>=0.74.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ty>=0.0.12", "apx==0.2.6",
]

[tool.apx.metadata]
app-name = "gainwell-genie-app"
app-slug = "gainwell_genie_app"
app-entrypoint = "gainwell_genie_app.backend.app:app"
api-prefix = "/api"
metadata-path = "gainwell_genie_app/_metadata.py"

[tool.apx.ui]
root = "gainwell_genie_app/ui"

[tool.apx.ui.registries]
"@animate-ui" = "https://animate-ui.com/r/{name}.json"
"@ai-elements" = "https://registry.ai-sdk.dev/{name}.json"
"@svgl" = "https://svgl.app/r/{name}.json"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
fallback-version = "0.0.0"

[tool.uv]
[[tool.uv.index]]
name = "apx-index"
url = "https://databricks-solutions.github.io/apx/simple"

[tool.uv.sources]
apx = { index = "apx-index" }

[tool.hatch.build.hooks.version]
path = "gainwell_genie_app/_version.py"
template = 'version = "{version}"'

[tool.hatch.build]
artifacts = ["gainwell_genie_app/__dist__", "gainwell_genie_app/_metadata.py"]
exclude = ["gainwell_genie_app/ui"]
